refactor(video-editor): split the canvas and drop the cut-area scrim
video_editor_canvas.dart had accumulated several self-contained pieces
that were impossible to test without pumping the whole editor. Pull the
playhead math, the clip preview, the setup placeholder, the hit-test
expander and the two clip enums into their own files so each can be
covered directly.

Remove the cut-area scrim in the same pass. It dimmed the band between
the target crop rect and the editor's render space, and dates from when
the clip preview switched between contain and cover depending on the
clip's aspect ratio. The preview is contain-only now, so the dimmed band
no longer marks anything the user can act on — it only greyed out layers
parked outside the crop. With the bars gone, the zoom matrix that kept
them aligned during pinch has no consumer either, so the notifier and
its editor callback go with them.

The render-space mapping itself is untouched: renderSize keeps the clip's
source aspect ratio (the native texture is stretched to its box, so any
other ratio would distort it) and the cover-fit into targetSize is what
makes the cropped frame fill the visible rect. HitTestExpander stays for
the same reason as before — the letterbox around targetSize still needs
to forward pinch and pan into the editor.
