+++ mobile/lib/screens/video_editor/video_editor_screen.dart
@@ -108,11 +108,6 @@ class _VideoEditorScreenState extends ConsumerState<VideoEditorScreen>
   /// Body size notifier, updated by [_CanvasFitter].
   final _bodySizeNotifier = ValueNotifier<Size>(Size.zero);
 
-  /// Editor zoom transform notifier (identity = not zoomed), driven by the
-  /// editor's zoom matrix. The letterbox scrim applies the same transform so
-  /// the bars move with the magnified frame.
-  final _zoomMatrixNotifier = ValueNotifier<Matrix4>(Matrix4.identity());
-
   /// Fine-grained editor play time (timeline space), updated by the canvas on
   /// every playhead tick so canvas overlays track playback smoothly.
   final _playTimeNotifier = ValueNotifier<Duration>(Duration.zero);
@@ -264,7 +259,6 @@ class _VideoEditorScreenState extends ConsumerState<VideoEditorScreen>
     _timelineOverlayBloc.close();
     _isLoadingDraft.dispose();
     _bodySizeNotifier.dispose();
-    _zoomMatrixNotifier.dispose();
     _playTimeNotifier.dispose();
     super.dispose();
   }
@@ -987,7 +981,6 @@ class _VideoEditorScreenState extends ConsumerState<VideoEditorScreen>
               removeAreaKey: _removeAreaKey,
               originalClipAspectRatio: clip?.originalAspectRatio ?? 9 / 16,
               bodySizeNotifier: _bodySizeNotifier,
-              zoomMatrixNotifier: _zoomMatrixNotifier,
               playTimeNotifier: _playTimeNotifier,
               fromLibrary: widget.fromLibrary,
               onOpenCamera: () => _openCamera(
