+++ mobile/test/widgets/video_editor/draw_editor/video_editor_draw_bottom_bar_test.dart
@@ -83,7 +83,6 @@ void main() {
                 removeAreaKey: GlobalKey(),
                 originalClipAspectRatio: 9 / 16,
                 bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-                zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                 playTimeNotifier: ValueNotifier(Duration.zero),
                 fromLibrary: false,
                 onOpenCamera: () {},
