+++ mobile/test/widgets/video_editor/filter_editor/video_editor_filter_bottom_bar_test.dart
@@ -75,7 +75,6 @@ void main() {
                 onAddEditTextLayer: ([_]) async => null,
                 originalClipAspectRatio: 9 / 16,
                 bodySizeNotifier: bodySizeNotifier,
-                zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                 playTimeNotifier: ValueNotifier(Duration.zero),
                 fromLibrary: false,
                 child: const Scaffold(body: VideoEditorFilterBottomBar()),
