+++ mobile/test/widgets/video_editor/filter_editor/video_editor_filter_overlay_controls_test.dart
@@ -70,7 +70,6 @@ void main() {
             removeAreaKey: GlobalKey(),
             originalClipAspectRatio: 9 / 16,
             bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-            zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
             playTimeNotifier: ValueNotifier(Duration.zero),
             fromLibrary: false,
             onOpenCamera: () {},
