+++ mobile/test/widgets/video_editor/main_editor/video_editor_main_actions_sheet_test.dart
@@ -340,7 +340,6 @@ VideoEditorScope _scope({
   onOpenCaptions: onOpenCaptions ?? () {},
   originalClipAspectRatio: 9 / 16,
   bodySizeNotifier: ValueNotifier(const Size(400, 800)),
-  zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
   playTimeNotifier: ValueNotifier(Duration.zero),
   fromLibrary: false,
   child: child ?? const SizedBox.shrink(),
