+++ mobile/test/widgets/video_editor/main_editor/video_editor_main_bottom_bar_test.dart
@@ -57,7 +57,6 @@ Widget _buildWidget({
       onOpenCaptions: () {},
       originalClipAspectRatio: 9 / 16,
       bodySizeNotifier: ValueNotifier(const Size(400, 800)),
-      zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
       playTimeNotifier: ValueNotifier(Duration.zero),
       fromLibrary: false,
       child: const Scaffold(body: VideoEditorMainBottomBar()),
