+++ mobile/test/widgets/video_editor/main_editor/video_editor_main_overlay_actions_test.dart
@@ -142,7 +142,6 @@ void main() {
                 removeAreaKey: GlobalKey(),
                 originalClipAspectRatio: 9 / 16,
                 bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-                zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                 playTimeNotifier: ValueNotifier(Duration.zero),
                 fromLibrary: false,
                 onOpenCamera: () {},
