+++ mobile/test/widgets/video_editor/main_editor/video_editor_remove_area_test.dart
@@ -47,7 +47,6 @@ void main() {
               removeAreaKey: GlobalKey(),
               originalClipAspectRatio: 9 / 16,
               bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-              zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
               playTimeNotifier: ValueNotifier(Duration.zero),
               fromLibrary: false,
               onOpenCamera: () {},
