+++ mobile/test/widgets/video_editor/main_editor/video_editor_scope_test.dart
@@ -51,7 +51,6 @@ void main() {
             onOpenCaptions: () {},
             originalClipAspectRatio: 9 / 16,
             bodySizeNotifier: ValueNotifier(const Size(400, 800)),
-            zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
             playTimeNotifier: ValueNotifier(Duration.zero),
             fromLibrary: false,
             child: Builder(
@@ -93,7 +92,6 @@ void main() {
             onOpenCaptions: () {},
             originalClipAspectRatio: 9 / 16,
             bodySizeNotifier: ValueNotifier(const Size(400, 800)),
-            zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
             playTimeNotifier: ValueNotifier(Duration.zero),
             fromLibrary: false,
             child: Builder(
