+++ mobile/test/widgets/video_editor/stop_motion/stop_motion_frame_commands_test.dart
@@ -104,7 +104,6 @@ void main() {
             removeAreaKey: GlobalKey(),
             originalClipAspectRatio: 9 / 16,
             bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-            zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
             playTimeNotifier: ValueNotifier(Duration.zero),
             fromLibrary: false,
             onOpenCamera: () {},
