+++ mobile/test/widgets/video_editor/timeline_editor/controls/video_editor_layer_animation_sheet_test.dart
@@ -511,7 +511,6 @@ void main() {
             removeAreaKey: GlobalKey(),
             originalClipAspectRatio: 9 / 16,
             bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-            zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
             playTimeNotifier: ValueNotifier(Duration.zero),
             fromLibrary: false,
             onOpenCamera: () {},
