+++ mobile/test/widgets/video_editor/timeline_editor/controls/video_editor_timeline_clip_controls_test.dart
@@ -133,7 +133,6 @@ void main() {
                 removeAreaKey: GlobalKey(),
                 originalClipAspectRatio: 9 / 16,
                 bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-                zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                 playTimeNotifier: ValueNotifier(Duration.zero),
                 fromLibrary: false,
                 onOpenCamera: () {},
@@ -575,7 +574,6 @@ void main() {
                   removeAreaKey: GlobalKey(),
                   originalClipAspectRatio: 9 / 16,
                   bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-                  zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                   playTimeNotifier: ValueNotifier(Duration.zero),
                   fromLibrary: false,
                   onOpenCamera: () {},
@@ -692,7 +690,6 @@ void main() {
                         removeAreaKey: GlobalKey(),
                         originalClipAspectRatio: 9 / 16,
                         bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-                        zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                         playTimeNotifier: ValueNotifier(Duration.zero),
                         fromLibrary: false,
                         onOpenCamera: () {},
