+++ mobile/test/widgets/video_editor/timeline_editor/controls/video_editor_timeline_layer_multi_select_controls_test.dart
@@ -64,7 +64,6 @@ void main() {
                 removeAreaKey: GlobalKey(),
                 originalClipAspectRatio: 9 / 16,
                 bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-                zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                 playTimeNotifier: ValueNotifier(Duration.zero),
                 fromLibrary: false,
                 onOpenCamera: () {},
