+++ mobile/test/widgets/video_editor/timeline_editor/controls/video_editor_timeline_overlay_controls_test.dart
@@ -81,7 +81,6 @@ void main() {
                 removeAreaKey: GlobalKey(),
                 originalClipAspectRatio: 9 / 16,
                 bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-                zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                 playTimeNotifier: ValueNotifier(Duration.zero),
                 fromLibrary: false,
                 onOpenCamera: () {},
@@ -111,7 +110,6 @@ void main() {
               removeAreaKey: GlobalKey(),
               originalClipAspectRatio: 9 / 16,
               bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-              zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
               playTimeNotifier: ValueNotifier(Duration.zero),
               fromLibrary: false,
               onOpenCamera: () {},
@@ -919,7 +917,6 @@ void main() {
                       removeAreaKey: GlobalKey(),
                       originalClipAspectRatio: 9 / 16,
                       bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-                      zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
                       playTimeNotifier: ValueNotifier(Duration.zero),
                       fromLibrary: false,
                       onOpenCamera: () {},
