+++ mobile/test/widgets/video_editor/timeline_editor/video_editor_timeline_test.dart
@@ -158,7 +158,6 @@ void main() {
               removeAreaKey: GlobalKey(),
               originalClipAspectRatio: 9 / 16,
               bodySizeNotifier: ValueNotifier(const Size(400, 600)),
-              zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
               playTimeNotifier: ValueNotifier(Duration.zero),
               fromLibrary: false,
               onOpenCamera: () {},
