+++ mobile/test/widgets/video_editor/video_editor_scaffold_test.dart
@@ -84,7 +84,6 @@ void main() {
           onOpenCaptions: () {},
           originalClipAspectRatio: 9 / 16,
           bodySizeNotifier: ValueNotifier(const Size(400, 800)),
-          zoomMatrixNotifier: ValueNotifier(Matrix4.identity()),
           playTimeNotifier: ValueNotifier(Duration.zero),
           fromLibrary: false,
           child: MultiBlocProvider(
